BROKEN COMMIT--DOES NOT COMPILE.
Modify (not fully) List functions to support the Stack ADT instead
of using malloc--have not tested. Let's address the elephant in
the room: the compile error. I have created some spaghetti code
by trying to declare a singleton(ish) Stack in stack.h and/or
list.h--this is causing build error, not sure if it's fixable
with the way I'm trying to do things (spaghetti way).