Add List_remove and List_prev (doesn't work).
I think the problem might be that I tried assigning the OOB
enums to pointers. I only did it because I didn't understand
what the hell they would be used for otherwise. Perhaps it's okay
to do, but maybe a better way would be better to give another property
to the Node struct, of type OOB, which acts as a flag. Having it
on the pointer itself just seems like, even if you get it to work,
it shouldn't. Not sure if having the enum flag will fix it.